Fix cookiesort()
The previous version didn't sort cookies correctly. We should not have to look
at path unless the domain of the two cookies is identical. We also shouldn't
have to look at the key unless domain and path of both cookies is identical.

If you want to tell me that compiler macros are ugly: Well, it works and the
alternative wouldn't be nice either.

Signed-off-by: Uli Schlachter <psychon@znc.in>